refactor: 커플 연결 단계 흐름 로직을 별도 훅으로 분리
